carburants_dbt/scripts/requirements.txt
requests>=2.31.0
snowflake-connector-python>=3.6.0
pandas>=2.0.0
